kmeans/requirements.txt
flask==2.1.0
pillow==*
numpy==*
opencv-python==*